third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/the-input-element/type-change-state-expected.txt

This is a testharness.js-based test.
Found 462 tests; 307 PASS, 155 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS change state from hidden to text
PASS change state from hidden to search
PASS change state from hidden to tel
PASS change state from hidden to url
PASS change state from hidden to email
PASS change state from hidden to password
PASS change state from hidden to datetime-local
PASS change state from hidden to date
PASS change state from hidden to month
PASS change state from hidden to week
PASS change state from hidden to time
PASS change state from hidden to number
PASS change state from hidden to range
PASS change state from hidden to color
PASS change state from hidden to checkbox
PASS change state from hidden to radio
PASS change state from hidden to file
PASS change state from hidden to submit
PASS change state from hidden to image
PASS change state from hidden to reset
PASS change state from hidden to button
FAIL change state from text to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from text to search
PASS change state from text to tel
PASS change state from text to url
PASS change state from text to email
PASS change state from text to password
PASS change state from text to datetime-local
PASS change state from text to date
PASS change state from text to month
PASS change state from text to week
PASS change state from text to time
PASS change state from text to number
PASS change state from text to range
PASS change state from text to color
FAIL change state from text to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from text to file
FAIL change state from text to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from search to text
PASS change state from search to tel
PASS change state from search to url
PASS change state from search to email
PASS change state from search to password
PASS change state from search to datetime-local
PASS change state from search to date
PASS change state from search to month
PASS change state from search to week
PASS change state from search to time
PASS change state from search to number
PASS change state from search to range
PASS change state from search to color
FAIL change state from search to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from search to file
FAIL change state from search to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from tel to text
PASS change state from tel to search
PASS change state from tel to url
PASS change state from tel to email
PASS change state from tel to password
PASS change state from tel to datetime-local
PASS change state from tel to date
PASS change state from tel to month
PASS change state from tel to week
PASS change state from tel to time
PASS change state from tel to number
PASS change state from tel to range
PASS change state from tel to color
FAIL change state from tel to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from tel to file
FAIL change state from tel to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from url to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from url to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
FAIL change state from url to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
FAIL change state from url to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
PASS change state from url to email
FAIL change state from url to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
PASS change state from url to datetime-local
PASS change state from url to date
PASS change state from url to month
PASS change state from url to week
PASS change state from url to time
PASS change state from url to number
PASS change state from url to range
PASS change state from url to color
FAIL change state from url to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from url to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
PASS change state from url to file
FAIL change state from url to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from url to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from url to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from url to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
FAIL change state from email to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
FAIL change state from email to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
PASS change state from email to url
FAIL change state from email to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
PASS change state from email to datetime-local
PASS change state from email to date
PASS change state from email to month
PASS change state from email to week
PASS change state from email to time
PASS change state from email to number
PASS change state from email to range
PASS change state from email to color
FAIL change state from email to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
PASS change state from email to file
FAIL change state from email to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from password to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from password to text
PASS change state from password to search
PASS change state from password to tel
PASS change state from password to url
PASS change state from password to email
PASS change state from password to datetime-local
PASS change state from password to date
PASS change state from password to month
PASS change state from password to week
PASS change state from password to time
PASS change state from password to number
PASS change state from password to range
PASS change state from password to color
FAIL change state from password to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from password to file
FAIL change state from password to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from datetime-local to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from datetime-local to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from datetime-local to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from datetime-local to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from datetime-local to urlassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from datetime-local to emailassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from datetime-local to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
PASS change state from datetime-local to date
PASS change state from datetime-local to month
PASS change state from datetime-local to week
PASS change state from datetime-local to time
PASS change state from datetime-local to number
PASS change state from datetime-local to range
PASS change state from datetime-local to color
FAIL change state from datetime-local to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from datetime-local to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
PASS change state from datetime-local to file
FAIL change state from datetime-local to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from datetime-local to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from datetime-local to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from datetime-local to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from date to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from date to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from date to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from date to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from date to urlassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from date to emailassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from date to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
PASS change state from date to datetime-local
PASS change state from date to month
PASS change state from date to week
PASS change state from date to time
PASS change state from date to number
PASS change state from date to range
PASS change state from date to color
FAIL change state from date to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from date to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
PASS change state from date to file
FAIL change state from date to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from date to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from date to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from date to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from month to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from month to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from month to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from month to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from month to urlassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from month to emailassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from month to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
PASS change state from month to datetime-local
PASS change state from month to date
PASS change state from month to week
PASS change state from month to time
PASS change state from month to number
PASS change state from month to range
PASS change state from month to color
FAIL change state from month to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from month to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
PASS change state from month to file
FAIL change state from month to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from month to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from month to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from month to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from week to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from week to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from week to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from week to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from week to urlassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from week to emailassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from week to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
PASS change state from week to datetime-local
PASS change state from week to date
PASS change state from week to month
PASS change state from week to time
PASS change state from week to number
PASS change state from week to range
PASS change state from week to color
FAIL change state from week to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from week to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
PASS change state from week to file
FAIL change state from week to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from week to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from week to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from week to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from time to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from time to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from time to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from time to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from time to urlassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from time to emailassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from time to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
PASS change state from time to datetime-local
PASS change state from time to date
PASS change state from time to month
PASS change state from time to week
PASS change state from time to number
PASS change state from time to range
PASS change state from time to color
FAIL change state from time to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from time to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
PASS change state from time to file
FAIL change state from time to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from time to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from time to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from time to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from number to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from number to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from number to urlassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from number to emailassert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from number to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
PASS change state from number to datetime-local
PASS change state from number to date
PASS change state from number to month
PASS change state from number to week
PASS change state from number to time
PASS change state from number to range
PASS change state from number to color
FAIL change state from number to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
PASS change state from number to file
FAIL change state from number to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from range to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "50"
FAIL change state from range to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "50"
FAIL change state from range to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "50"
FAIL change state from range to urlassert_equals: input.value should be foobar after change of state expected "foobar" but got "50"
FAIL change state from range to emailassert_equals: input.value should be foobar after change of state expected "foobar" but got "50"
FAIL change state from range to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "50"
PASS change state from range to datetime-local
PASS change state from range to date
PASS change state from range to month
PASS change state from range to week
PASS change state from range to time
FAIL change state from range to numberassert_equals: input.value should be  after change of state expected "" but got "50"
PASS change state from range to color
FAIL change state from range to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
PASS change state from range to file
FAIL change state from range to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from color to hiddenassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "#000000"
FAIL change state from color to textassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "#000000"
FAIL change state from color to searchassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "#000000"
FAIL change state from color to telassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "#000000"
FAIL change state from color to urlassert_equals: input.value should be foobar after change of state expected "foobar" but got "#000000"
FAIL change state from color to emailassert_equals: input.value should be foobar after change of state expected "foobar" but got "#000000"
FAIL change state from color to passwordassert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "#000000"
PASS change state from color to datetime-local
PASS change state from color to date
PASS change state from color to month
PASS change state from color to week
PASS change state from color to time
PASS change state from color to number
PASS change state from color to range
FAIL change state from color to checkboxassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "#000000"
FAIL change state from color to radioassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "#000000"
PASS change state from color to file
FAIL change state from color to submitassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "#000000"
FAIL change state from color to imageassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "#000000"
FAIL change state from color to resetassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "#000000"
FAIL change state from color to buttonassert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "#000000"
PASS change state from checkbox to hidden
PASS change state from checkbox to text
PASS change state from checkbox to search
PASS change state from checkbox to tel
PASS change state from checkbox to url
PASS change state from checkbox to email
PASS change state from checkbox to password
PASS change state from checkbox to datetime-local
PASS change state from checkbox to date
PASS change state from checkbox to month
PASS change state from checkbox to week
PASS change state from checkbox to time
PASS change state from checkbox to number
PASS change state from checkbox to range
PASS change state from checkbox to color
PASS change state from checkbox to radio
PASS change state from checkbox to file
PASS change state from checkbox to submit
PASS change state from checkbox to image
PASS change state from checkbox to reset
PASS change state from checkbox to button
PASS change state from radio to hidden
PASS change state from radio to text
PASS change state from radio to search
PASS change state from radio to tel
PASS change state from radio to url
PASS change state from radio to email
PASS change state from radio to password
PASS change state from radio to datetime-local
PASS change state from radio to date
PASS change state from radio to month
PASS change state from radio to week
PASS change state from radio to time
PASS change state from radio to number
PASS change state from radio to range
PASS change state from radio to color
PASS change state from radio to checkbox
PASS change state from radio to file
PASS change state from radio to submit
PASS change state from radio to image
PASS change state from radio to reset
PASS change state from radio to button
PASS change state from file to hidden
PASS change state from file to text
PASS change state from file to search
PASS change state from file to tel
PASS change state from file to url
PASS change state from file to email
PASS change state from file to password
PASS change state from file to datetime-local
PASS change state from file to date
PASS change state from file to month
PASS change state from file to week
PASS change state from file to time
PASS change state from file to number
PASS change state from file to range
PASS change state from file to color
PASS change state from file to checkbox
PASS change state from file to radio
PASS change state from file to submit
PASS change state from file to image
PASS change state from file to reset
PASS change state from file to button
PASS change state from submit to hidden
PASS change state from submit to text
PASS change state from submit to search
PASS change state from submit to tel
PASS change state from submit to url
PASS change state from submit to email
PASS change state from submit to password
PASS change state from submit to datetime-local
PASS change state from submit to date
PASS change state from submit to month
PASS change state from submit to week
PASS change state from submit to time
PASS change state from submit to number
PASS change state from submit to range
PASS change state from submit to color
PASS change state from submit to checkbox
PASS change state from submit to radio
PASS change state from submit to file
PASS change state from submit to image
PASS change state from submit to reset
PASS change state from submit to button
PASS change state from image to hidden
PASS change state from image to text
PASS change state from image to search
PASS change state from image to tel
PASS change state from image to url
PASS change state from image to email
PASS change state from image to password
PASS change state from image to datetime-local
PASS change state from image to date
PASS change state from image to month
PASS change state from image to week
PASS change state from image to time
PASS change state from image to number
PASS change state from image to range
PASS change state from image to color
PASS change state from image to checkbox
PASS change state from image to radio
PASS change state from image to file
PASS change state from image to submit
PASS change state from image to reset
PASS change state from image to button
PASS change state from reset to hidden
PASS change state from reset to text
PASS change state from reset to search
PASS change state from reset to tel
PASS change state from reset to url
PASS change state from reset to email
PASS change state from reset to password
PASS change state from reset to datetime-local
PASS change state from reset to date
PASS change state from reset to month
PASS change state from reset to week
PASS change state from reset to time
PASS change state from reset to number
PASS change state from reset to range
PASS change state from reset to color
PASS change state from reset to checkbox
PASS change state from reset to radio
PASS change state from reset to file
PASS change state from reset to submit
PASS change state from reset to image
PASS change state from reset to button
PASS change state from button to hidden
PASS change state from button to text
PASS change state from button to search
PASS change state from button to tel
PASS change state from button to url
PASS change state from button to email
PASS change state from button to password
PASS change state from button to datetime-local
PASS change state from button to date
PASS change state from button to month
PASS change state from button to week
PASS change state from button to time
PASS change state from button to number
PASS change state from button to range
PASS change state from button to color
PASS change state from button to checkbox
PASS change state from button to radio
PASS change state from button to file
PASS change state from button to submit
PASS change state from button to image
PASS change state from button to reset
Harness: the test ran to completion.
